requirements.txt
jupyter
pyaaas
matplotlib
locustio